Completed Full Stack Todo App

+++ TodoAppBackend/src/main/java/com/yifan/demo/service/TodoHardcodedService.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Service
 public class TodoHardcodedService {
     private static List<Todo> todos = new ArrayList();
-    private static int idCounter = 0;
+    private static long idCounter = 0;
 
     static {
         todos.add(new Todo(++idCounter,"yifan","Learn to dance",new Date(), false));
